srcs/requirements/backend/files/requirements.txt
# Django
Django==4.2.7
# Postgre DB Adapter
psycopg2-binary==2.9.9
# Rest Framework
djangorestframework==3.14.0
# CORS (maybe needed for Rest-API)
# django-cors-headers==4.3.1